Created ProductPage and moved context to wrap the whole application. Fetching products on initial mount
