+++ ecommerce-client/src/components/DisplayProducts.tsx
@@ -1,23 +1,11 @@
-import { useContext, useEffect } from "react"
+import { useContext } from "react"
 import ProductCard from "./ProductCard"
-import { ProductContext } from "../contexts/ProductContext"
-import { ActionType } from "../reducers/CustomerReducer";
-import { useProduct } from "../hooks/useProducts";
+import ProductContext from "../contexts/ProductContext";
 
 export const DisplayProducts = () => {
-    const { fetchProductsHandler } = useProduct()
-    const { products, dispatch } = useContext(ProductContext);
 
-    useEffect(() => {
-        const getData = async () => {
-            const data = await fetchProductsHandler();
-            dispatch({
-                type: ActionType.LOADED,
-                payload: JSON.stringify(data),
-            });
-        };
-        getData();
-    }, []);
+    const { products } = useContext(ProductContext)
+    console.log("Display products", products)
 
     return (
         <>
